fif(app): fix comments for PR
