+++ .eslintrc.json
@@ -41,7 +41,9 @@
       "extends": [
         "plugin:@angular-eslint/template/recommended"
       ],
-      "rules": {}
+      "rules": {
+        "semi": ["error", "always"]
+      }
     }
   ]
 }
