--- src/app/shared/customer-button/customer-button.component.html
@@ -1,3 +1,0 @@
-<button type="btn" [disabled]="isDisabled" (click)="clickHandler()" [class]="classExpression">
-    <span>{{title}}</span>
-</button>
