+++ src/app/shared/index.ts
@@ -1,0 +1,1 @@
+export { TodoButtonComponent } from './todo-button/todo-button.component';
