+++ src/app/shared/todo-button/todo-button.component.html
@@ -1,0 +1,3 @@
+<button type="btn" [disabled]="isDisabled" (click)="clickHandler()" [class]="BtnColor">
+    <span>{{title}}</span>
+</button>
